+++ libkwave/undo/UndoDeleteAction.h
@@ -58,10 +58,10 @@ namespace Kwave
 	virtual QString description();
 
 	/** @see UndoAction::undoSize() */
-	virtual unsigned int undoSize();
+	virtual qint64 undoSize();
 
 	/** @see UndoAction::redoSize() */
-	virtual int redoSize();
+	virtual qint64 redoSize();
 
 	/**
 	 * Stores the data needed for undo.
